+++ certbot-dns-nsone/certbot_dns_nsone/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_nsone.dns_nsone` plugin.
