+++ certbot-dns-nsone/docs/api/dns_nsone.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_nsone.dns_nsone`
+:mod:`certbot_dns_nsone._internal.dns_nsone`
 ----------------------------------
 
-.. automodule:: certbot_dns_nsone.dns_nsone
+.. automodule:: certbot_dns_nsone._internal.dns_nsone
    :members:
